requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
halftone==0.0.1
    # via -r requirements.in
numpy==1.25.1
    # via halftone
pillow==10.0.0
    # via halftone
